Refactor to use Responses and Renderers.
